Cleaned up code for opensea fetching and tokenhistory
